Create an instrument to generate stings
