+++ test/jest-component/features/steps/events/post-events-steps.ts
@@ -1,11 +1,15 @@
 import { StepDefinitions } from 'jest-cucumber';
-import { eventRepository, context } from '../../../steps-config';
+import { eventRepository, context, schema } from '../../../steps-config';
 import { CREATE_EVENT_MOCK } from '../../../../mocks/dtos';
+import { GenRand } from '../../../instruments/gen-random';
+
+const genRand = new GenRand();
 
 export const postEventsSteps: StepDefinitions = ({ given, and }) => {
   given('the POST Events API is available', () => {
     context.path = '/v1/events';
     context.mock = { ...CREATE_EVENT_MOCK };
+    context.method = 'post';
   });
 
   and('an unauthenticated user', () => {});
@@ -13,11 +17,46 @@ export const postEventsSteps: StepDefinitions = ({ given, and }) => {
   and('an authenticated user with admin rights is logged in', () => {});
   and(
     'user creates a new event with the maximum allowed characters in all input fields',
-    () => {},
+    () => {
+      for (const row in schema.paths[context.path][context.method].requestBody
+        .content['application/json'].schema.properties) {
+        const prop =
+          schema.paths[context.path][context.method].requestBody.content[
+            'application/json'
+          ].schema.properties[row];
+        if (prop.type == 'string' && prop.maxLength) {
+          const data_length = prop.maxLength;
+          if (row.split('.').length == 2) {
+            context.mock[row.split('.')[0]][row.split('.')[1]] =
+              genRand.genRandString(data_length);
+          } else {
+            context.mock[row] = genRand.genRandString(data_length);
+          }
+        }
+      }
+    },
   );
   and(
     'user creates a new event with the minimum allowed characters in all input fields',
-    () => {},
+    () => {
+      for (const row in schema.paths[context.path][context.method].requestBody
+        .content['application/json'].schema.properties) {
+        const prop =
+          schema.paths[context.path][context.method].requestBody.content[
+            'application/json'
+          ].schema.properties[row];
+
+        if (prop.type == 'string' && prop.minLength) {
+          const data_length = prop.minLength;
+          if (row.split('.').length == 2) {
+            context.mock[row.split('.')[0]][row.split('.')[1]] =
+              genRand.genRandString(data_length);
+          } else {
+            context.mock[row] = genRand.genRandString(data_length);
+          }
+        }
+      }
+    },
   );
   and('with valid data to create an event', () => {});
   and(
