+++ test/jest-component/instruments/gen-random.ts
@@ -1,0 +1,15 @@
+export class GenRand {
+  constructor() {}
+  genRandString(length: number): string {
+    const chars =
+      'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+    let result = '';
+    const charsLength = chars.length;
+
+    for (let i = 0; i < length; i++) {
+      const randomIndex = Math.floor(Math.random() * charsLength);
+      result += chars[randomIndex];
+    }
+    return result;
+  }
+}
